chore: PR review updates part 1
