+++ bc_obps/compliance/service/compliance_dashboard_service.py
@@ -6,14 +6,8 @@ from service.data_access_service.user_service import UserDataAccessService
 from service.data_access_service.operation_service import OperationDataAccessService
 from registration.models.operation import Operation
 from typing import Optional
-from dataclasses import dataclass
 from compliance.service.elicensing.elicensing_data_refresh_service import ElicensingDataRefreshService
-
-
-@dataclass
-class PaymentDataWithFreshnessFlag:
-    data_is_fresh: bool
-    data: QuerySet[ElicensingPayment]
+from compliance.dataclass import PaymentDataWithFreshnessFlag
 
 
 class ComplianceDashboardService:
